Scope spaces to teams
With this change, all spaces MUST be assigned to a team.

NOTE: for now, the sidebar contains the full list of both teams and spaces. This will change as we add users to teams and introduce permissions and access restrictions (e.g. this may change to show only the current user's teams and spaces).